Split the LockFileFormat
Moving the methods responsible of writing into
Microsoft.Framework.PackageManager so that json serializing is not
required in Microsoft.Framework.Runtime
