+++ src/Microsoft.Framework.Runtime/DependencyManagement/LockFileFormat.cs
@@ -1,14 +1,14 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.Versioning;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using System;
-using System.IO;
-using System.Collections.Generic;
 using NuGet;
-using System.Runtime.Versioning;
-using System.Linq;
 
 namespace Microsoft.Framework.Runtime.DependencyManagement
 {
@@ -25,7 +25,7 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             }
         }
 
-        public LockFile Read(Stream stream)
+        private LockFile Read(Stream stream)
         {
             using (var textReader = new StreamReader(stream))
             {
@@ -56,28 +56,6 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             }
         }
 
-        public void Write(string filePath, LockFile lockFile)
-        {
-            using (var stream = new FileStream(filePath, FileMode.Create, FileAccess.Write, FileShare.None))
-            {
-                Write(stream, lockFile);
-            }
-        }
-
-        public void Write(Stream stream, LockFile lockFile)
-        {
-            using (var textWriter = new StreamWriter(stream))
-            {
-                using (var jsonWriter = new JsonTextWriter(textWriter))
-                {
-                    jsonWriter.Formatting = Formatting.Indented;
-
-                    var json = WriteLockFile(lockFile);
-                    json.WriteTo(jsonWriter);
-                }
-            }
-        }
-
         private LockFile ReadLockFile(JObject cursor)
         {
             var lockFile = new LockFile();
@@ -87,17 +65,6 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             lockFile.Targets = ReadObject(cursor["targets"] as JObject, ReadTarget);
             lockFile.ProjectFileDependencyGroups = ReadObject(cursor["projectFileDependencyGroups"] as JObject, ReadProjectFileDependencyGroup);
             return lockFile;
-        }
-
-        private JObject WriteLockFile(LockFile lockFile)
-        {
-            var json = new JObject();
-            json["locked"] = new JValue(lockFile.Islocked);
-            json["version"] = new JValue(Version);
-            json["targets"] = WriteObject(lockFile.Targets, WriteTarget);
-            json["libraries"] = WriteObject(lockFile.Libraries, WriteLibrary);
-            json["projectFileDependencyGroups"] = WriteObject(lockFile.ProjectFileDependencyGroups, WriteProjectFileDependencyGroup);
-            return json;
         }
 
         private LockFileLibrary ReadLibrary(string property, JToken json)
@@ -113,29 +80,6 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             library.Sha512 = ReadString(json["sha512"]);
             library.Files = ReadPathArray(json["files"] as JArray, ReadString);
             return library;
-        }
-
-        private JProperty WriteLibrary(LockFileLibrary library)
-        {
-            var json = new JObject();
-            if (library.IsServiceable)
-            {
-                WriteBool(json, "serviceable", library.IsServiceable);
-            }
-            json["sha512"] = WriteString(library.Sha512);
-            WritePathArray(json, "files", library.Files, WriteString);
-            return new JProperty(
-                library.Name + "/" + library.Version.ToString(),
-                json);
-        }
-
-        private JProperty WriteTarget(LockFileTarget target)
-        {
-            var json = WriteObject(target.Libraries, WriteTargetLibrary);
-
-            var key = target.TargetFramework + (target.RuntimeIdentifier == null ? "" : "/" + target.RuntimeIdentifier);
-
-            return new JProperty(key, json);
         }
 
         private LockFileTarget ReadTarget(string property, JToken json)
@@ -173,50 +117,11 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             return library;
         }
 
-        private JProperty WriteTargetLibrary(LockFileTargetLibrary library)
-        {
-            var json = new JObject();
-
-            if (library.Dependencies.Count > 0)
-            {
-                json["dependencies"] = WriteObject(library.Dependencies, WritePackageDependency);
-            }
-
-            if (library.FrameworkAssemblies.Count > 0)
-            {
-                json["frameworkAssemblies"] = WriteArray(library.FrameworkAssemblies, WriteFrameworkAssemblyReference);
-            }
-
-            if (library.CompileTimeAssemblies.Count > 0)
-            {
-                json["compile"] = WritePathArray(library.CompileTimeAssemblies, WriteString);
-            }
-
-            if (library.RuntimeAssemblies.Count > 0)
-            {
-                json["runtime"] = WritePathArray(library.RuntimeAssemblies, WriteString);
-            }
-
-            if (library.NativeLibraries.Count > 0)
-            {
-                json["native"] = WritePathArray(library.NativeLibraries, WriteString);
-            }
-
-            return new JProperty(library.Name + "/" + library.Version, json);
-        }
-
         private ProjectFileDependencyGroup ReadProjectFileDependencyGroup(string property, JToken json)
         {
             return new ProjectFileDependencyGroup(
                 property,
                 ReadArray(json as JArray, ReadString));
-        }
-
-        private JProperty WriteProjectFileDependencyGroup(ProjectFileDependencyGroup frameworkInfo)
-        {
-            return new JProperty(
-                frameworkInfo.FrameworkName,
-                WriteArray(frameworkInfo.Dependencies, WriteString));
         }
 
         private IList<FrameworkAssemblyReference> ReadFrameworkAssemblies(JObject json)
@@ -241,34 +146,6 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             }).ToList();
         }
 
-        private void WriteFrameworkAssemblies(JToken json, string property, IList<FrameworkAssemblyReference> frameworkAssemblies)
-        {
-            if (frameworkAssemblies.Any())
-            {
-                json[property] = WriteFrameworkAssemblies(frameworkAssemblies);
-            }
-        }
-
-        private JToken WriteFrameworkAssemblies(IList<FrameworkAssemblyReference> frameworkAssemblies)
-        {
-            var groups = frameworkAssemblies.SelectMany(x =>
-            {
-                if (x.SupportedFrameworks.Any())
-                {
-                    return x.SupportedFrameworks.Select(xx => new { x.AssemblyName, FrameworkName = xx });
-                }
-                else
-                {
-                    return new[] { new { x.AssemblyName, FrameworkName = default(FrameworkName) } };
-                }
-            }).GroupBy(x => x.FrameworkName);
-
-            return WriteObject(groups, group =>
-            {
-                return new JProperty(group.Key.ToStringSafe() ?? "*", group.Select(x => new JValue(x.AssemblyName)));
-            });
-        }
-
         private PackageDependencySet ReadPackageDependencySet(string property, JToken json)
         {
             var targetFramework = string.Equals(property, "*") ? null : new FrameworkName(property);
@@ -276,14 +153,6 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
                 targetFramework,
                 ReadObject(json as JObject, ReadPackageDependency));
         }
-
-        private JProperty WritePackageDependencySet(PackageDependencySet item)
-        {
-            return new JProperty(
-                item.TargetFramework.ToStringSafe() ?? "*",
-                WriteObject(item.Dependencies, WritePackageDependency));
-        }
-
 
         private PackageDependency ReadPackageDependency(string property, JToken json)
         {
@@ -293,21 +162,9 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
                 versionStr == null ? null : VersionUtility.ParseVersionSpec(versionStr));
         }
 
-        private JProperty WritePackageDependency(PackageDependency item)
-        {
-            return new JProperty(
-                item.Id,
-                WriteString(item.VersionSpec.ToStringSafe()));
-        }
-
         private FrameworkAssemblyReference ReadFrameworkAssemblyReference(JToken json)
         {
             return new FrameworkAssemblyReference(json.Value<string>());
-        }
-
-        private JToken WriteFrameworkAssemblyReference(FrameworkAssemblyReference item)
-        {
-            return new JValue(item.AssemblyName);
         }
 
         private PackageReferenceSet ReadPackageReferenceSet(JToken json)
@@ -318,25 +175,11 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
                 ReadArray(json["references"] as JArray, ReadString));
         }
 
-        private JToken WritePackageReferenceSet(PackageReferenceSet item)
-        {
-            var json = new JObject();
-            json["targetFramework"] = item.TargetFramework.ToStringSafe();
-            json["references"] = WriteArray(item.References, WriteString);
-            return json;
-        }
-
         private IPackageFile ReadPackageFile(string property, JToken json)
         {
             var file = new LockFilePackageFile();
             file.Path = PathUtility.GetPathWithDirectorySeparator(property);
             return file;
-        }
-
-        private JProperty WritePackageFile(IPackageFile item)
-        {
-            var json = new JObject();
-            return new JProperty(PathUtility.GetPathWithForwardSlashes(item.Path), new JObject());
         }
 
         private IList<TItem> ReadArray<TItem>(JArray json, Func<JToken, TItem> readItem)
@@ -358,34 +201,6 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             return ReadArray(json, readItem).Select(f => PathUtility.GetPathWithDirectorySeparator(f)).ToList();
         }
 
-        private void WriteArray<TItem>(JToken json, string property, IEnumerable<TItem> items, Func<TItem, JToken> writeItem)
-        {
-            if (items.Any())
-            {
-                json[property] = WriteArray(items, writeItem);
-            }
-        }
-
-        private void WritePathArray(JToken json, string property, IEnumerable<string> items, Func<string, JToken> writeItem)
-        {
-            WriteArray(json, property, items.Select(f => PathUtility.GetPathWithForwardSlashes(f)), writeItem);
-        }
-
-        private JArray WriteArray<TItem>(IEnumerable<TItem> items, Func<TItem, JToken> writeItem)
-        {
-            var array = new JArray();
-            foreach (var item in items)
-            {
-                array.Add(writeItem(item));
-            }
-            return array;
-        }
-
-        private JArray WritePathArray(IEnumerable<string> items, Func<string, JToken> writeItem)
-        {
-            return WriteArray(items.Select(f => PathUtility.GetPathWithForwardSlashes(f)), writeItem);
-        }
-
         private IList<TItem> ReadObject<TItem>(JObject json, Func<string, JToken, TItem> readItem)
         {
             if (json == null)
@@ -398,24 +213,6 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
                 items.Add(readItem(child.Key, child.Value));
             }
             return items;
-        }
-
-        private void WriteObject<TItem>(JToken json, string property, IEnumerable<TItem> items, Func<TItem, JProperty> writeItem)
-        {
-            if (items.Any())
-            {
-                json[property] = WriteObject(items, writeItem);
-            }
-        }
-
-        private JObject WriteObject<TItem>(IEnumerable<TItem> items, Func<TItem, JProperty> writeItem)
-        {
-            var array = new JObject();
-            foreach (var item in items)
-            {
-                array.Add(writeItem(item));
-            }
-            return array;
         }
 
         private bool ReadBool(JToken cursor, string property, bool defaultValue)
@@ -453,23 +250,9 @@ namespace Microsoft.Framework.Runtime.DependencyManagement
             return SemanticVersion.Parse(valueToken.Value<string>());
         }
 
-        private void WriteBool(JToken token, string property, bool value)
-        {
-            token[property] = new JValue(value);
-        }
-
-        private JToken WriteString(string item)
-        {
-            return item != null ? new JValue(item) : JValue.CreateNull();
-        }
-
         private FrameworkName ReadFrameworkName(JToken json)
         {
             return json == null ? null : new FrameworkName(json.Value<string>());
-        }
-        private JToken WriteFrameworkName(FrameworkName item)
-        {
-            return item != null ? new JValue(item.ToString()) : JValue.CreateNull();
         }
 
         class LockFilePackageFile : IPackageFile
